+++ ThirdPartyAdapters/yahoo/yahoo/src/main/java/com/google/ads/mediation/yahoo/YahooRewardedRenderer.java
@@ -1,7 +1,7 @@
-package com.google.ads.mediation.verizon;
+package com.google.ads.mediation.yahoo;
 
-import static com.google.ads.mediation.verizon.VerizonMediationAdapter.TAG;
-import static com.google.ads.mediation.verizon.VerizonMediationAdapter.initializeSDK;
+import static com.google.ads.mediation.yahoo.YahooMediationAdapter.TAG;
+import static com.google.ads.mediation.yahoo.YahooMediationAdapter.initializeSDK;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -13,16 +13,20 @@ import com.google.android.gms.ads.mediation.MediationRewardedAd;
 import com.google.android.gms.ads.mediation.MediationRewardedAdCallback;
 import com.google.android.gms.ads.mediation.MediationRewardedAdConfiguration;
 import com.google.android.gms.ads.rewarded.RewardItem;
-import com.verizon.ads.ErrorInfo;
-import com.verizon.ads.interstitialplacement.InterstitialAd;
-import com.verizon.ads.interstitialplacement.InterstitialAdFactory;
-import com.verizon.ads.utils.ThreadUtils;
+import com.yahoo.ads.ErrorInfo;
+import com.yahoo.ads.interstitialplacement.InterstitialAd;
+import com.yahoo.ads.interstitialplacement.InterstitialAd.InterstitialAdListener;
+import com.yahoo.ads.interstitialplacement.InterstitialPlacementConfig;
+import com.yahoo.ads.utils.ThreadUtils;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListener,
-    InterstitialAdFactory.InterstitialAdFactoryListener, MediationRewardedAd {
+final class YahooRewardedRenderer implements InterstitialAdListener,
+    MediationRewardedAd {
 
+  /**
+   * Yahoo Mobile SDK specific event ID for video completion.
+   */
   private static final String VIDEO_COMPLETE_EVENT_ID = "onVideoComplete";
 
   /**
@@ -32,7 +36,7 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
       mediationAdLoadCallback;
 
   /**
-   * Verizon Media rewarded ad.
+   * Yahoo rewarded ad.
    */
   private InterstitialAd rewardedAd;
 
@@ -51,48 +55,65 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
    */
   private final MediationRewardedAdConfiguration mediationRewardedAdConfiguration;
 
-  public VerizonMediaRewardedRenderer(@NonNull MediationAdLoadCallback<MediationRewardedAd,
-      MediationRewardedAdCallback> mediationAdLoadCallback, MediationRewardedAdConfiguration
-      mediationRewardedAdConfiguration) {
+  public YahooRewardedRenderer(
+      @NonNull MediationRewardedAdConfiguration mediationRewardedAdConfiguration,
+      @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
+          mediationAdLoadCallback) {
     this.mediationAdLoadCallback = mediationAdLoadCallback;
     this.mediationRewardedAdConfiguration = mediationRewardedAdConfiguration;
   }
 
   public void render() {
     Bundle serverParameters = mediationRewardedAdConfiguration.getServerParameters();
-    String siteId = VerizonMediaAdapterUtils.getSiteId(serverParameters,
+    String siteId = YahooAdapterUtils.getSiteId(serverParameters,
         mediationRewardedAdConfiguration);
     Context context = mediationRewardedAdConfiguration.getContext();
     if (!initializeSDK(context, siteId)) {
-      final String message = "Unable to initialize Verizon Ads SDK.";
+      final String message = "Unable to initialize the Yahoo Mobile SDK.";
       Log.e(TAG, message);
       mediationAdLoadCallback.onFailure(message);
       return;
     }
 
-    String placementId = VerizonMediaAdapterUtils.getPlacementId(serverParameters);
+    String placementId = YahooAdapterUtils.getPlacementId(serverParameters);
     if (TextUtils.isEmpty(placementId)) {
-      mediationAdLoadCallback.onFailure(
-          "Verizon Ads SDK placement ID must be set in mediationRewardedAdConfiguration" +
-              " server params.");
+      mediationAdLoadCallback.onFailure("Yahoo Mobile SDK placement ID must be set in "
+          + "mediationRewardedAdConfiguration server params.");
       return;
     }
 
-    VerizonMediaAdapterUtils.setCoppaValue(mediationRewardedAdConfiguration);
-    InterstitialAdFactory interstitialAdFactory =
-        new InterstitialAdFactory(mediationRewardedAdConfiguration.getContext(),
-            placementId, this);
-    interstitialAdFactory.setRequestMetaData(VerizonMediaAdapterUtils
-        .getRequestMetaData(mediationRewardedAdConfiguration));
-    interstitialAdFactory.load(this);
+    YahooAdapterUtils.setCoppaValue(mediationRewardedAdConfiguration);
+
+    InterstitialPlacementConfig placementConfig = new InterstitialPlacementConfig(placementId,
+        YahooAdapterUtils.getRequestMetaData(mediationRewardedAdConfiguration));
+    rewardedAd = new InterstitialAd(context, placementId, YahooRewardedRenderer.this);
+    rewardedAd.load(placementConfig);
   }
 
   @Override
-  public void onLoaded(final InterstitialAdFactory interstitialAdFactory,
-      final InterstitialAd interstitialAd) {
+  public void showAd(@NonNull Context context) {
+    if (rewardedAd == null) {
+      if (mediationRewardedAdCallback != null) {
+        mediationRewardedAdCallback.onAdFailedToShow("No ads to show.");
+      }
+      return;
+    }
+    rewardedAd.show(context);
+  }
 
-    Log.i(TAG, "Verizon Ads SDK incentivized video interstitial loaded.");
+  void destroy() {
+    if (rewardedAd != null) {
+      rewardedAd.destroy();
+    }
+  }
+
+  // region Yahoo InterstitialAdListener implementation.
+
+  @Override
+  public void onLoaded(final InterstitialAd interstitialAd) {
+    Log.i(TAG, "Yahoo Mobile SDK loaded a rewarded ad successfully.");
     this.rewardedAd = interstitialAd;
+
     // Reset the completion event with each new interstitial ad load.
     completionEventCalled.set(false);
     ThreadUtils.postOnUiThread(new Runnable() {
@@ -100,17 +121,16 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
       public void run() {
         if (mediationAdLoadCallback != null) {
           mediationRewardedAdCallback =
-              mediationAdLoadCallback.onSuccess(VerizonMediaRewardedRenderer.this);
+              mediationAdLoadCallback.onSuccess(YahooRewardedRenderer.this);
         }
       }
     });
   }
 
   @Override
-  public void onError(final InterstitialAdFactory interstitialAdFactory,
-      final ErrorInfo errorInfo) {
-    final String message = "Verizon Ads SDK incentivized video interstitial request failed (" +
-        errorInfo.getErrorCode() + "): " + errorInfo.getDescription();
+  public void onLoadFailed(final InterstitialAd interstitialAd, final ErrorInfo errorInfo) {
+    final String message = "Yahoo Mobile SDK failed to request a rewarded ad with error code: "
+        + errorInfo.getErrorCode() + ", message: " + errorInfo.getDescription();
     Log.w(TAG, message);
     ThreadUtils.postOnUiThread(new Runnable() {
       @Override
@@ -124,8 +144,7 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
 
   @Override
   public void onError(final InterstitialAd interstitialAd, final ErrorInfo errorInfo) {
-
-    Log.e(TAG, "Verizon Ads SDK incentivized video interstitial error: " + errorInfo);
+    Log.w(TAG, "Yahoo Mobile SDK returned an error for rewarded ad: " + errorInfo);
 
     // This error callback is used if the interstitial ad is loaded successfully, but an
     // error occurs while trying to display
@@ -141,7 +160,7 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
 
   @Override
   public void onShown(final InterstitialAd interstitialAd) {
-    Log.i(TAG, "Verizon Ads SDK incentivized video interstitial shown.");
+    Log.i(TAG, "Yahoo Mobile SDK showed a rewarded ad.");
     ThreadUtils.postOnUiThread(new Runnable() {
       @Override
       public void run() {
@@ -155,7 +174,7 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
 
   @Override
   public void onClosed(final InterstitialAd interstitialAd) {
-    Log.i(TAG, "Verizon Ads SDK ad closed.");
+    Log.i(TAG, "Yahoo Mobile SDK closed a rewarded ad.");
     ThreadUtils.postOnUiThread(new Runnable() {
       @Override
       public void run() {
@@ -168,7 +187,7 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
 
   @Override
   public void onClicked(final InterstitialAd interstitialAd) {
-    Log.i(TAG, "Verizon Ads SDK incentivized video interstitial clicked.");
+    Log.i(TAG, "Yahoo Mobile SDK recoded a click on a rewarded ad.");
     ThreadUtils.postOnUiThread(new Runnable() {
       @Override
       public void run() {
@@ -181,7 +200,8 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
 
   @Override
   public void onAdLeftApplication(final InterstitialAd interstitialAd) {
-    Log.i(TAG, "Verizon Ads SDK incentivized video interstitial left application.");
+    Log.i(TAG, "Yahoo Mobile SDK has caused the user to leave the application "
+        + "from a rewarded ad.");
   }
 
   @Override
@@ -189,6 +209,7 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
       final String eventId, final Map<String, Object> arguments) {
 
     if (VIDEO_COMPLETE_EVENT_ID.equals(eventId) && !completionEventCalled.getAndSet(true)) {
+      Log.i(TAG, "Yahoo Mobile SDK has completed playing a rewarded ad.");
       ThreadUtils.postOnUiThread(new Runnable() {
         @Override
         public void run() {
@@ -212,20 +233,5 @@ class VerizonMediaRewardedRenderer implements InterstitialAd.InterstitialAdListe
     }
   }
 
-  @Override
-  public void showAd(@NonNull Context context) {
-    if (rewardedAd == null) {
-      if (mediationRewardedAdCallback != null) {
-        mediationRewardedAdCallback.onAdFailedToShow("No ads to show.");
-      }
-      return;
-    }
-    rewardedAd.show(context);
-  }
-
-  void destroy() {
-    if (rewardedAd != null) {
-      rewardedAd.destroy();
-    }
-  }
+  // endregion
 }